Fix seed implementation
Let the old seed functions be backwards compatible,
and the new behave better in the functional api.

Remove the new part of seed/1 because the semantics was
so different between the old and new functionality.

Allow seed/2 to produce a new seed from a 3-tuple or from an old
state as seed/3 did in the previous implementation.

The new seed/2 returns the new state and not the old as seed/3 did.